requirements.txt
asgiref==3.7.2
autopep8==2.3.2
Django==5.0.13
pycodestyle==2.13.0
python-dotenv==1.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
firebase-functions
firebase-admin
gunicorn
psycopg2-binary
Pillow
pycountry
django-modeltranslation
